Move alarms converter code to a separate file
